Rename objects subpackage to entities
